added generated data
things changed from the original, credit score is now tied to a real number which varies, added field for how much they invest the month of when the data point was taken, and how much risk they are willing to take.  Note, this data is up for further adjustment when needed due to me lacking sufficient processing power.